Update interactive import.
